% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_cd}
\alias{benchmark_cd}
\title{Test my characteristic direction processing of l1000 data}
\usage{
benchmark_cd(dprimes, cd_es = NULL, l1000_es = NULL)
}
\description{
Test my characteristic direction processing of l1000 data
}
