% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_chdir}
\alias{benchmark_chdir}
\title{Test characteristic direction authors original processing of l1000 data}
\usage{
benchmark_chdir(dprimes, l1000_es = NULL, chdir_es = NULL)
}
\description{
Test characteristic direction authors original processing of l1000 data
}
