% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_diff_cds}
\alias{benchmark_diff_cds}
\title{Compare diff_cds to diff_expr processed query signatures}
\usage{
benchmark_diff_cds(anals_cd, dprimes, cd_es, l1000_es)
}
\description{
Compare diff_cds to diff_expr processed query signatures
}
