% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_ngenes}
\alias{benchmark_ngenes}
\title{Compare different number of selected genes}
\usage{
benchmark_ngenes(dprimes, drug_info, data_dir = getwd(), suffix = "_ngenes")
}
\description{
Compare different number of selected genes
}
