% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_cd.R
\name{diff_cds}
\alias{diff_cds}
\title{Characteristic Direction Differential Expression}
\usage{
diff_cds(esets, data_dir = getwd(), annot = "SYMBOL", prev_anals = list(NULL))
}
\description{
Characteristic Direction Differential Expression
}
