% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pvals}
\alias{get_pvals}
\title{Extract Adjusted P-Values}
\usage{
get_pvals(es, anals)
}
\arguments{
\item{es}{result of call to \code{es_meta}.}

\item{anals}{result of call to \code{diff_expr)}.}
}
\value{
List with slots:
  \item{meta}{Named numeric vector with overall false discovery rates for all gene from meta-analysis.}
  \item{contrasts}{List of named numeric vectors (one per contrast) with limma adjusted p-values for all genes.}
}
\description{
Used to benchmark p-value cutoffs for query signatures.
}
\seealso{
\code{\link[crossmeta]{es_meta}}, \code{\link[crossmeta]{diff_expr}}.
}
