% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_cd.R
\name{load_diff_cds}
\alias{load_diff_cds}
\title{Load Characteristic Direction Differential Expression Signatures}
\usage{
load_diff_cds(gse_names, data_dir = getwd(), annot = "SYMBOL")
}
\description{
Load Characteristic Direction Differential Expression Signatures
}
