% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_cor}
\alias{query_cor}
\title{Run Correlation Query}
\usage{
query_cor(query_genes, drug_info, ngenes = 100, method = "pearson")
}
\arguments{
\item{query_genes}{Numeric vector of effect size values with genes as names.}

\item{drug_info}{Matrix of effect size values to query against. Rows names are genes, column names are signature names.}

\item{ngenes}{Number of top upregulated and number of top downregulated \code{query_genes} that are in
common with genes in \code{drug_info}.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}
}
\value{
Vector of query values with names of corresponding signature in \code{drug_info} sorted by decreasing similarity.
}
\description{
Run Correlation Query
}
