% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_euclidean}
\alias{query_euclidean}
\title{Run Euclidean Distance Query}
\usage{
query_euclidean(query_genes, drug_info, ngenes = 100)
}
\arguments{
\item{query_genes}{Numeric vector of effect size values with genes as names.}

\item{drug_info}{Matrix of effect size values to query against. Rows names are genes, column names are signature names.}

\item{ngenes}{Number of top upregulated and number of top downregulated \code{query_genes} that are in
common with genes in \code{drug_info}.}
}
\value{
Vector of query values with names of corresponding signature in \code{drug_info} sorted by decreasing similarity.
}
\description{
Run Euclidean Distance Query
}
