% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ngenes.R
\name{query_fixed_ngenes}
\alias{query_fixed_ngenes}
\title{Query using fixed number of genes}
\usage{
query_fixed_ngenes(dprimes, drug_info, breaks = c(seq(50, 950, 50), 1001))
}
\arguments{
\item{dprimes}{List with names of signatures to query with. Each item of list is named a numeric vector of
effect size values with names corresponding to the gene.}

\item{drug_info}{Matrix of differential expression values to query against.
Row names are genes and column names are signatures.}

\item{breaks}{Breaks to use for determing \code{ngenes} argument to \code{query_drugs}.}
}
\value{
Named list with items:
 \item{resl}{For each value of \code{breaks}, a list of pearson correlations between each signature in
 \code{dprimes} and \code{drug_info}.}
 \item{ngenes}{For each value of \code{breaks}, a numeric vector indicating the number of genes used for the query.}


 @seealso \code{\link[ccmap]{query_drugs}}
}
\description{
Query using fixed number of genes
}
