% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_ks}
\alias{query_ks}
\title{Run KS Query}
\usage{
query_ks(query_genes, drug_prl, ngenes = 100)
}
\arguments{
\item{query_genes}{Numeric vector of effect size values with genes as names.}

\item{drug_prl}{Matrix of PRLs to query against. Rows names are genes, column names are signature names.}

\item{ngenes}{Number of top upregulated and number of top downregulated \code{query_genes} that are in
common with genes in \code{drug_info}.}
}
\value{
Vector of query values with names of corresponding signature in \code{drug_info} sorted by decreasing similarity.
}
\description{
Run KS Query
}
