% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ngenes.R
\name{query_pval_ngenes}
\alias{query_pval_ngenes}
\title{Query using p-values to choose number of genes}
\usage{
query_pval_ngenes(
  dprimes,
  drug_info,
  pvals,
  breaks = c(seq(0.01, 0.3, 0.01), seq(0.3, 1, 0.05))
)
}
\arguments{
\item{dprimes}{List with names of signatures to query with. Each item of list is named a numeric vector of
effect size values with names corresponding to the gene.}

\item{drug_info}{Matrix of differential expression values to query against.
Row names are genes and column names are signatures.}

\item{pvals}{Same as \code{dprimes} but p-values or FDRs in place of effect size values.}

\item{breaks}{Breaks to use for determing \code{ngenes} argument to \code{query_drugs}.}
}
\value{
Named list with items:
 \item{resl}{For each value of \code{breaks}, a list of pearson correlations between each signature in
 \code{dprimes} and \code{drug_info}.}
 \item{ngenes}{For each value of \code{breaks}, a numeric vector indicating the number of genes used for the query.}


 @seealso \code{\link[ccmap]{query_drugs}}
}
\description{
Query using p-values to choose number of genes
}
